# src/docs/swagger.yaml
basePath: /api/v1
definitions:
  api.ATCatalogResponse:
    properties:
      commands:
        items:
          $ref: '#/definitions/modem.ATCommandInfo'
        type: array
      supported:
        description: |-
          Supported lists the command names the modem reported through AT+CLAC,
          or is null when the modem does not support that query.
        items:
          type: string
        type: array
    type: object
...
  models.ATCommandRequest:
    properties:
      command:
        type: string
      confirm:
        description: Confirm must be true to run a dangerous or unrecognised command.
        type: boolean
    type: object
  models.ATCommandResponse:
    properties:
      response:
        type: string
    type: object
  models.ATConfirmationRequired:
    properties:
      error:
        type: string
      requires_confirmation:
        type: boolean
      risk:
        description: Risk is "dangerous" or "unknown".
        type: string
      title:
        description: Title names the recognised command, when there is one.
        type: string
      warning:
        type: string
    type: object
...
  models.Contact:
    properties:
      created_at:
        type: string
      name:
        type: string
      phone_number:
        type: string
      updated_at:
        type: string
    type: object
  models.ContactRequest:
    properties:
      name:
        example: Jane Doe
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.ModemInfoResponse:
    properties:
      firmware:
        type: string
      iccid:
        type: string
      imei:
        type: string
      imsi:
        type: string
      manufacturer:
        type: string
      model:
        type: string
      network:
        description: Network is the operator the modem is registered on, e.g. "T-Mobile".
        type: string
      phone_number:
        type: string
      provider:
        description: Provider is the service provider name stored on the SIM, e.g.
          "Tello".
        type: string
    type: object
...
  models.ModemSignalResponse:
    properties:
      quality:
        type: string
      signal:
        type: integer
    type: object
  models.ModemStatusResponse:
    properties:
      status:
        type: string
    type: object
...
  modem.ATCommandInfo:
    properties:
      category:
        type: string
      description:
        type: string
      example:
        type: string
      forms:
        items:
          $ref: '#/definitions/modem.ATForm'
        type: array
      name:
        description: |-
          Name is the command as written after "AT": "+CSQ", "&F", "I", or ""
          for the bare AT attention command.
        type: string
      params:
        items:
          $ref: '#/definitions/modem.ATParam'
        type: array
      reference:
        type: string
      response:
        type: string
      risk:
        allOf:
        - $ref: '#/definitions/modem.ATRisk'
        description: |-
          Risk applies to the execute and set forms. Read and test forms never
          change anything and are always safe.
      title:
        type: string
      warning:
        type: string
    type: object
  modem.ATForm:
    properties:
      description:
        type: string
      kind:
        $ref: '#/definitions/modem.ATFormKind'
      syntax:
        type: string
    type: object
  modem.ATFormKind:
    enum:
    - execute
    - read
    - test
    - set
    type: string
    x-enum-comments:
      FormExecute: AT+CMD
      FormRead: AT+CMD?
      FormSet: AT+CMD=<args>, or ATE0 for basic commands
      FormTest: AT+CMD=?
    x-enum-descriptions:
    - AT+CMD
    - AT+CMD?
    - AT+CMD=?
    - AT+CMD=<args>, or ATE0 for basic commands
    x-enum-varnames:
    - FormExecute
    - FormRead
    - FormTest
    - FormSet
  modem.ATParam:
    properties:
      description:
        type: string
      name:
        type: string
      values:
        items:
          $ref: '#/definitions/modem.ATParamValue'
        type: array
    type: object
  modem.ATParamValue:
    properties:
      description:
        type: string
      value:
        type: string
    type: object
  modem.ATRisk:
    enum:
    - safe
    - config
    - dangerous
    - unknown
    type: string
    x-enum-varnames:
    - RiskSafe
    - RiskConfig
    - RiskDangerous
    - RiskUnknown
host: localhost:5174
info:
  contact: {}
  description: REST API for sending and receiving SMS messages via a USB GSM modem.
  license:
    name: GPL-3.0
  title: SMS Gateway API
  version: "1.0"
paths:
...
  /api/v1/contacts:
    delete:
      description: Removes the saved name for a phone number. Messages are not affected.
      parameters:
      - description: Phone number of the contact
        in: query
        name: phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Delete contact name
      tags:
      - Contacts
    get:
      description: Returns saved contact names ordered by name.
      parameters:
      - description: Only contacts whose name or number contains this text
        in: query
        name: q
        type: string
      - description: Maximum contacts to return (max 500). Omit to return all.
        in: query
        name: limit
        type: integer
      - description: Contacts to skip. Only applied together with limit.
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Total matching contacts is returned in the X-Total-Count header
          schema:
            items:
              $ref: '#/definitions/models.Contact'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: List contacts
      tags:
      - Contacts
    put:
      consumes:
      - application/json
      description: Sets the display name for a phone number, creating the contact
        if needed.
      parameters:
      - description: Phone number of the contact
        in: query
        name: phone
        required: true
        type: string
      - description: Contact name
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ContactRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Contact'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Save contact name
      tags:
      - Contacts
...
  /api/v1/modem/at:
    post:
      consumes:
      - application/json
      description: Sends a raw AT command to the GSM modem and returns the response.
        Requires admin privileges.
      parameters:
      - description: AT command to send
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ATCommandRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ATCommandResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Dangerous or unrecognised command sent without confirm
          schema:
            $ref: '#/definitions/models.ATConfirmationRequired'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Send AT command
      tags:
      - Modem
  /api/v1/modem/at/commands:
    get:
      description: Returns documentation for common AT commands, with the risk of
        running each, and the command names the modem reports through AT+CLAC (null
        when unsupported). Requires admin privileges.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ATCatalogResponse'
      security:
      - BearerAuth: []
      summary: AT command reference
      tags:
      - Modem
  /api/v1/modem/info:
    get:
      description: Returns the line provider and network, the SIM phone number, ICCID
        and IMSI, and the modem IMEI, manufacturer, model and firmware. A field is
        empty when the modem or SIM cannot report it; many SIMs do not store their
        phone number. The details are read once and cached until the service restarts;
        pass refresh=true to read them from the modem again. Requires admin privileges.
      parameters:
      - description: Read the details from the modem again instead of using the cached
          ones
        in: query
        name: refresh
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ModemInfoResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get modem and SIM details
      tags:
      - Modem
...
  /api/v1/modem/signal:
    get:
      description: Returns the GSM modem signal strength and quality assessment.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ModemSignalResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get modem signal strength
      tags:
      - Modem
  /api/v1/modem/status:
    get:
      description: Checks whether the GSM modem is responsive and returns its current
        status.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ModemStatusResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/models.ModemStatusResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get modem status
      tags:
      - Modem
...
securityDefinitions:
  ApiKeyAuth:
    description: API key for programmatic access
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: JWT token with Bearer prefix (e.g. "Bearer eyJhbG...")
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"